layers: syncing viewports based on active layers because pan and rotate wasnt working
Every time a non-base layer was being moved and drawComposition was called
we were updating the rotation and translation of the active layer based
on the base layer and when doing this the current viewport values were
being lost resulting and the image wasnt being rotated/translated.
That's because when a layer is set as active, enabledElement.viewport
is updated to the active one.